+++ changelog.md
@@ -1,4 +1,4 @@
-# Changes in 1.0.1
+# Changes in 1.1.0
 
 ## Breaking changes